move Navbar component using Next.js's private folders
