+++ app/layout.tsx
@@ -1,12 +1,6 @@
-import type { Metadata } from 'next'
 import { ClerkProvider } from '@clerk/nextjs'
 import './globals.css'
-import Navbar from './components/Navbar'
-
-export const metadata: Metadata = {
-  title: 'HMS',
-  description: 'Hockey Management System',
-}
+import Navbar from './_components/Navbar'
 
 export default function RootLayout({
   children,
@@ -16,6 +10,11 @@ export default function RootLayout({
   return (
     <ClerkProvider>
       <html lang="en">
+        <head>
+          <title>Goal Guardian</title>
+          {/* TODO: Generate an actual description */}
+          <meta name="description" content="Goal Guardian is a hockey team management system" />
+        </head>
         <body>
           <Navbar />
